Uploading cleaned-up version of our code
Hi everybody! Thanks for letting me join the group.

As I worked to get up to speed with all that's been done this semester, it seemed to me a service I could offer the group was a fairly naive set of eyes to check the code's human readability. 

As I read the code, I took the liberty of making a couple (mostly aesthetic) changes to bring the code in line with the usual python style guide. I'm uploading the product of that work here. It includes the code up to and including everything we did today. In general, these are the changes I made:
1: Moved working bits of code into functions and wrote descriptions of the functions
2: Deleted (most) small or incomplete annotations in favor of string descriptions of functions
3: Created (commented) headings for imports, definitions, global variables, function calls, and a space at the bottom for experimental stuff.

In addition, I made a slight tweak to the parsing function we defined today (all documented in the comments, and put some thoughts into a comment at the bottom about how we might modify the (currently working) stack to work in a more iterative fashion, based on my current level of understanding.